Add function to list collected subscriptions in SQLite
The main change in this commit is the addition of a function to list collected subscriptions in the SQLite database. It also includes a restructuring of several other parts of the code mainly involving database connections, as a function to instantiate a connection to the CostManagementStore has been abstracted out and is now used in various places instead of repeating the same block of code.